// src/integration/kube/resources/object-meta.ts
// SPDX-License-Identifier: Apache-2.0

import {type NamespaceName} from '../../../types/namespace/namespace-name.js';

export interface ObjectMeta {
  readonly namespace?: NamespaceName;
  readonly name: string;
  readonly labels?: {[key: string]: string};
  readonly annotations?: {[key: string]: string};
  readonly uid?: string;
}